third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-024-expected.txt


FAIL CSS Test: @viewport constrained - used height is positive. assert_equals: expected 1 but got 600
